requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
matplotlib==3.8.2
